// src/app/models/unit.ts
export interface Unit {
        id?: string;
        name?: string;
        lat?: string;
        lng?: string;
        added?: string;
        locality?: string;
        description?: string;
        province?: string;
        open?: string;
        phone?: string;
        photo?:any;
        updated?:any;
}